GeoServerRest interface extracted
